refactor: Relocate config builders to a dedicated directory and introduce a base builder class.
